Create ci-cd.yml
Create ci-cd.yml
Tiffany's example.
Important note! Since the beginning of the course, there has been significant development of the cd job in ci-cd.yml. Please copy over a new version of the file from here, and remove the {% raw %} and {% endraw %} tags (they are there to tell the Cookiecutter to not modify the code between them), as well as replace the {{ cookiecutter.THING }} templates with the corresponding value for your  pycounts* package. (edited) 

https://github.com/ttimbers/pycounts_tt_2024/blob/main/.github/workflows/ci-cd.yml